+++ CHANGELOG.md
@@ -5,6 +5,10 @@
 * Return a subtype of `B2Exception` on errors when the response body does not conform to `B2ErrorStructure`. 
   Returning a `B2Exception` subtype enables the `B2Retryer` to retry exceptions that may succeed on retry.
 
+### Fixed
+* Fixed B2ListFilesIterableBase assuming a response with 0 results was the end. It now looks for
+  `nextFileName` being null to indicate the end.
+
 ## [6.1.0] - 2022-09-19
 ### Added
 * Added support for Java 8's `-parameters` option so constructor parameters do not need to be reiterated in `B2Json.constructor#params`